requirements.txt
Django>=5.1,<5.2
django-environ==0.11.2
gunicorn==22.0.0
psycopg[binary]==3.2.1
whitenoise==6.7.0
Pillow==10.4.0
google-api-python-client==2.146.0